requirements.txt
anyio==4.3.0
asgiref==3.8.1
channels==4.1.0
click==8.1.7
colorama==0.4.6
Django==4.2.12
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
h11==0.14.0
httptools==0.6.1
idna==3.7
# matlabengine==23.2.3
matlabengine==24.1.2
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
plotly==5.20.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-libsbml==5.20.2
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tenacity==8.2.3
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
